backend/requirements.txt
# FastAPI and ASGI
fastapi
uvicorn[standard]
gunicorn

# Database
sqlalchemy
psycopg2-binary
alembic

# Authentication & Security
python-jose[cryptography]
passlib[bcrypt]
python-multipart

# HTTP Client & External APIs
httpx
requests

# Real-time Communication
python-socketio
python-socketio[asyncio_client]

# Data Validation & Serialization
pydantic
pydantic-settings
email-validator

# File Upload
aiofiles

# Environment & Configuration
python-dotenv

# AI/ML
opencv-python
numpy
pandas
scikit-learn
joblib
ultralytics

# Notification Services
firebase-admin
twilio

# CORS
fastapi-cors

# Logging & Monitoring
structlog

# Testing
pytest
pytest-asyncio

# Utilities
python-dateutil
Pillow
slowapi
